прошелся по комментам, всё сделал

+++ app/src/main/java/com/example/projectexcursions/ui/excursion/ExcursionActivity.kt
@@ -32,7 +32,7 @@ class ExcursionActivity : AppCompatActivity() {
 
     private fun initData() {
         val excursionId = intent.getLongExtra(EXTRA_EXCURSION_ID, -1)
-        if (excursionId == (-1).toLong()) {
+        if (excursionId == -1L) {
             Toast.makeText(this, "Invalidnaya exursia", Toast.LENGTH_SHORT).show()
             finish()
             return
